+++ lib/servers/data_vault/test/test_server.py
@@ -100,7 +100,7 @@ class DataVaultTest(unittest.TestCase):
         msg_head = 'Mismatch in entry {} of data.'
         for i, (expected_entry, actual_entry) in enumerate(comparables):
             msg = msg_head.format(i) + msg_foot
-            if isinstance(expected_entry, (int, long, float, complex)):
+            if isinstance(expected_entry, (int, int, float, complex)):
                 self.assertEqual(expected_entry, actual_entry, msg=msg)
             else:
                 self.assertArrayEqual(expected_entry, actual_entry, msg=msg)
